Start moving session functionality to sessions.go
